+++ node/src/mempool/persist.rs
@@ -356,6 +356,7 @@ mod tests {
             4,
             Default::default(),
             Default::default(),
+            Default::default(),
         )
         .unwrap()
     }