+++ node/src/mempool/pool.rs
@@ -3135,6 +3135,7 @@ mod tests {
             4,
             Default::default(),
             Default::default(),
+            Default::default(),
         )
         .unwrap();
         let mp = Mempool::new(1_000_000, 0); // 1MB, no min fee for tests
@@ -3254,6 +3255,7 @@ mod tests {
             AssumeValid::Disabled,
             450,
             4,
+            Default::default(),
             Default::default(),
             Default::default(),
         )
@@ -4437,6 +4439,7 @@ mod tests {
             AssumeValid::Disabled,
             450,
             4,
+            Default::default(),
             Default::default(),
             Default::default(),
         )