+++ node/src/mining/miner.rs
@@ -255,7 +255,8 @@ mod tests {
             450,
         4,
         Default::default(),
-        Default::default(),)
+        Default::default(),
+            Default::default(),)
         .unwrap();
         let mp = Mempool::new(1_000_000, 0);
 
@@ -287,7 +288,8 @@ mod tests {
             450,
         4,
         Default::default(),
-        Default::default(),)
+        Default::default(),
+            Default::default(),)
         .unwrap();
         let mp = Mempool::new(1_000_000, 0);
 
@@ -320,7 +322,8 @@ mod tests {
             450,
         4,
         Default::default(),
-        Default::default(),)
+        Default::default(),
+            Default::default(),)
         .unwrap();
         let mp = Mempool::new(1_000_000, 0);
 
@@ -351,7 +354,8 @@ mod tests {
             450,
         4,
         Default::default(),
-        Default::default(),)
+        Default::default(),
+            Default::default(),)
         .unwrap();
         let mp = Mempool::new(1_000_000, 0);
 
@@ -386,7 +390,8 @@ mod tests {
             450,
         4,
         Default::default(),
-        Default::default(),)
+        Default::default(),
+            Default::default(),)
         .unwrap();
         let mp = Mempool::new(1_000_000, 0);